requirements.txt
requests>=2.28.0
gspread>=6.0.0
google-auth>=2.0.0
pandas>=2.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pytz>=2023.3
fastapi>=0.100.0
uvicorn>=0.22.0
pytest>=7.0.0
